+++ lib/edgedb/types/relative_duration.ex
@@ -4,7 +4,7 @@ defmodule EdgeDB.RelativeDuration do
 
   ```elixir
   iex(1)> {:ok, pid} = EdgeDB.start_link()
-  iex(2)> EdgeDB.query_required_single!(pid, "SELECT <cal::relative_duration>'45.6 seconds'")
+  iex(2)> EdgeDB.query_required_single!(pid, "select <cal::relative_duration>'45.6 seconds'")
   #EdgeDB.RelativeDuration<"PT45.6S">
   ```
   """