+++ test/edgedb/protocol/codecs/config_memory_test.exs
@@ -7,21 +7,21 @@ defmodule Tests.EdgeDB.Protocol.Codecs.ConfigMemoryTest do
     value = 42 * 1024 * 1024
 
     assert %EdgeDB.ConfigMemory{bytes: ^value} =
-             EdgeDB.query_single!(conn, "SELECT <cfg::memory>'42MiB'")
+             EdgeDB.query_single!(conn, "select <cfg::memory>'42MiB'")
   end
 
   test "encoding integer as cfg::memory value", %{conn: conn} do
     value = 42
 
     assert %EdgeDB.ConfigMemory{bytes: ^value} =
-             EdgeDB.query_single!(conn, "SELECT <cfg::memory>$0", [value])
+             EdgeDB.query_single!(conn, "select <cfg::memory>$0", [value])
   end
 
   test "encoding EdgeDB.ConfigMemory as cfg::memory value", %{conn: conn} do
     value = %EdgeDB.ConfigMemory{bytes: 42 * 1024}
 
     assert true ==
-             EdgeDB.query_single!(conn, "SELECT <cfg::memory>$0 = <cfg::memory>'42KiB'", [value])
+             EdgeDB.query_single!(conn, "select <cfg::memory>$0 = <cfg::memory>'42KiB'", [value])
   end
 
   test "error when passing too large number as cfg::memory argument", %{conn: conn} do
@@ -29,7 +29,7 @@ defmodule Tests.EdgeDB.Protocol.Codecs.ConfigMemoryTest do
 
     exc =
       assert_raise EdgeDB.Error, fn ->
-        EdgeDB.query_single!(conn, "SELECT <cfg::memory>$0", [value])
+        EdgeDB.query_single!(conn, "select <cfg::memory>$0", [value])
       end
 
     assert exc ==
@@ -43,7 +43,7 @@ defmodule Tests.EdgeDB.Protocol.Codecs.ConfigMemoryTest do
 
     exc =
       assert_raise EdgeDB.Error, fn ->
-        EdgeDB.query_single!(conn, "SELECT <cfg::memory>$0", [value])
+        EdgeDB.query_single!(conn, "select <cfg::memory>$0", [value])
       end
 
     assert exc ==