+++ test/edgedb/protocol/codecs/local_date_test.exs
@@ -6,12 +6,12 @@ defmodule Tests.EdgeDB.Protocol.Codecs.LocalDateTest do
   test "decoding cal::local_date value", %{conn: conn} do
     value = ~D[2019-05-06]
 
-    assert ^value = EdgeDB.query_single!(conn, "SELECT <cal::local_date>'2019-05-06'")
+    assert ^value = EdgeDB.query_single!(conn, "select <cal::local_date>'2019-05-06'")
   end
 
   test "encoding cal::local_date value", %{conn: conn} do
     value = ~D[2019-05-06]
 
-    assert ^value = EdgeDB.query_single!(conn, "SELECT <cal::local_date>$0", [value])
+    assert ^value = EdgeDB.query_single!(conn, "select <cal::local_date>$0", [value])
   end
 end