+++ test/edgedb/protocol/codecs/local_time_test.exs
@@ -6,12 +6,12 @@ defmodule Tests.EdgeDB.Protocol.Codecs.LocalTimeTest do
   test "decoding cal::local_time value", %{conn: conn} do
     value = ~T[12:10:00]
 
-    assert ^value = EdgeDB.query_single!(conn, "SELECT <cal::local_time>'12:10'")
+    assert ^value = EdgeDB.query_single!(conn, "select <cal::local_time>'12:10'")
   end
 
   test "encoding cal::local_time value", %{conn: conn} do
     value = ~T[12:10:00]
 
-    assert ^value = EdgeDB.query_single!(conn, "SELECT <cal::local_time>$0", [value])
+    assert ^value = EdgeDB.query_single!(conn, "select <cal::local_time>$0", [value])
   end
 end